Fixed an issue with deploying to the root context in the openshift maven profile. Removed all old tests and tweaked the arquillian tests.
