+++ helloworld-ws/openshift/WEB-INF/jboss-web.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jboss-web>
+    <context-root>/</context-root>
+</jboss-web>
+
