--- helloworld-ws/src/test/java/org/jboss/as/quickstarts/wshelloworld/ArquillianClientTest.java
@@ -1,96 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2012, Red Hat, Inc. and/or its affiliates, and individual
- * contributors by the @authors tag. See the copyright.txt in the 
- * distribution for a full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,  
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.as.quickstarts.wshelloworld;
-
-import static org.junit.Assert.*;
-
-import java.io.File;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
-
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.as.quickstarts.wshelloworld.HelloWorldService;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * Simple set of tests for the HelloWorld Web Service to demonstrate accessing the web service using a client
- * 
- * @author lnewson@redhat.com
- */
-@RunWith(Arquillian.class)
-public class ArquillianClientTest {
-    /**
-     * The location of the WebApp source folder so we know where to find the web.xml when deploying using Arquillian.
-     */
-    private static final String WEBAPP_SRC = "src/main/webapp";
-    /**
-     * The name of the WAR Archive that will be used by Arquillian to deploy the application.
-     */
-    private static final String APP_NAME = "jboss-as-helloworld-ws";
-    /**
-     * The path of the WSDL endpoint in relation to the deployed web application.
-     */
-    private static final String WSDL_PATH = "HelloWorldService?wsdl";
-
-    @ArquillianResource
-    private URL deploymentUrl;
-
-    private HelloWorldService client;
-
-    @Deployment(testable = false)
-    public static WebArchive createDeployment() {
-        return ShrinkWrap.create(WebArchive.class, APP_NAME + ".war").addPackage(HelloWorldService.class.getPackage())
-                .addAsWebInfResource(new File(WEBAPP_SRC, "WEB-INF/web.xml"));
-    }
-
-    @Before
-    public void setup() {
-        try {
-            client = new Client(new URL(deploymentUrl, WSDL_PATH));
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Test
-    public void testHello() {
-        assertEquals(client.sayHello(), "Hello World!");
-    }
-
-    @Test
-    public void testHelloName() {
-        assertEquals(client.sayHelloToName("John"), "Hello John!");
-    }
-
-    @Test
-    public void testHelloNames() {
-        final List<String> names = new ArrayList<String>();
-        names.add("John");
-        names.add("Mary");
-        names.add("Mark");
-
-        assertEquals(client.sayHelloToNames(names), "Hello John, Mary & Mark!");
-    }
-}
