+++ helloworld-ws/src/test/java/org/jboss/as/quickstarts/wshelloworld/ClientTest.java
@@ -18,41 +18,57 @@ package org.jboss.as.quickstarts.wshelloworld;
 
 import static org.junit.Assert.*;
 
+import java.io.File;
 import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.as.quickstarts.wshelloworld.HelloWorldService;
-import org.junit.BeforeClass;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * Simple set of tests for the HelloWorld Web Service to demonstrate accessing the web service using a client
  * 
- * Note: These tests assume a JBoss AS7 server is already running on the local machine.
- * 
  * @author lnewson@redhat.com
  */
+@RunWith(Arquillian.class)
 public class ClientTest {
     /**
-     * The name of the WAR Archive that will be when deploying the application.
+     * The location of the WebApp source folder so we know where to find the web.xml when deploying using Arquillian.
+     */
+    private static final String WEBAPP_SRC = "src/main/webapp";
+    /**
+     * The name of the WAR Archive that will be used by Arquillian to deploy the application.
      */
     private static final String APP_NAME = "jboss-as-helloworld-ws";
     /**
-     * The base URL of the application when it's running.
+     * The path of the WSDL endpoint in relation to the deployed web application.
      */
-    private static final String BASE_URL = "http://localhost:8080/" + APP_NAME;
-    /**
-     * The url to the Web Service endpoint.
-     */
-    private static final String WSDL_URL = BASE_URL + "/HelloWorldService?wsdl";
+    private static final String WSDL_PATH = "HelloWorldService?wsdl";
 
-    private static HelloWorldService client;
+    @ArquillianResource
+    private URL deploymentUrl;
 
-    @BeforeClass
-    public static void setup() {
+    private HelloWorldService client;
+
+    @Deployment(testable = false)
+    public static WebArchive createDeployment() {
+        return ShrinkWrap.create(WebArchive.class, APP_NAME + ".war").addPackage(HelloWorldService.class.getPackage())
+                .addAsWebInfResource(new File(WEBAPP_SRC, "WEB-INF/web.xml"));
+    }
+
+    @Before
+    public void setup() {
         try {
-            client = new Client(WSDL_URL);
+            client = new Client(new URL(deploymentUrl, WSDL_PATH));
         } catch (MalformedURLException e) {
             e.printStackTrace();
         }
